Wstęp do dodania menu i game_over

+++ flappy_bird.py
@@ -2,10 +2,12 @@
 from pipe import Pipe
 from Bird import Bird
 from pipe import Pipe
+from game_state import State
 import random
 import arcade
 import os
 from game_variables import *
+# from buttons import *
 
 class Game(arcade.Window):
 
@@ -20,12 +22,15 @@ class Game(arcade.Window):
         self.background = None
         #List of birds
         self.bird_list = None
+        #Initial state of game
+        self.state = State.MAIN_MENU
 
         self.flapped = False
         self.score = 0
         self.dx = 0 # how many pixels pipes move
 
     def setup(self):
+        
         self.background = arcade.load_texture(BACKGROUNDS[0])
         self.bird_list = arcade.SpriteList()
         self.pipe_sprites = arcade.SpriteList()
@@ -37,6 +42,9 @@ class Game(arcade.Window):
         # Create starts pipes
         start_pipes = Pipe.random_size_pipe(self.height,self.width)
         self.pipe_sprites.extend(start_pipes)
+
+        # #Create buttons
+        # self.play_again = PlayAgainButton(self.width // 2 , self.height //3)
         
     def draw(self):
         """Funcion draw background"""
@@ -44,7 +52,7 @@ class Game(arcade.Window):
 
     def on_draw(self):
         """ This is the method called when the drawing time comes."""
-
+        self.ui_manager = None
         #render all objects
         arcade.start_render()
 
@@ -54,39 +62,47 @@ class Game(arcade.Window):
 
     def on_key_release(self, symbol, modifiers):
         if symbol == arcade.key.SPACE:
-            self.flapped = True
+            if self.state == State.PLAYING:
+                self.flapped = True
+            elif self.state == State.MAIN_MENU:
+                self.state = State.PLAYING
 
     def update(self,dt):
-        
-        if self.flapped:
-            self.bird.flap()
-            self.flapped = False
+        if self.state == State.PLAYING:
+            if self.flapped:
+                self.bird.flap()
+                self.flapped = False
 
-        # Check if bird is too high
-        if self.bird.top > self.height:
-            self.bird.top = self.height
+            # Check if bird is too high
+            if self.bird.top > self.height:
+                self.bird.top = self.height
 
-        # Check if bird is too low
-        if self.bird.bottom <= 0:
-            self.bird.bottom = 0
-        
-        self.dx += PIPE_SPEED
-        if self.dx > random.randrange(MIN_DISTACE_OF_PIPES, MAX_DISTACE_OF_PIPES):
-            self.dx = 0
-            new_pipe = Pipe.random_size_pipe(self.height, self.width)
-            self.pipe_sprites.extend(new_pipe)
+            # Check if bird is too low
+            if self.bird.bottom <= 0:
+                self.bird.bottom = 0
+            
+            self.dx += PIPE_SPEED
+            if self.dx > random.randrange(MIN_DISTACE_OF_PIPES, MAX_DISTACE_OF_PIPES):
+                self.dx = 0
+                new_pipe = Pipe.random_size_pipe(self.height, self.width)
+                self.pipe_sprites.extend(new_pipe)
 
-        for pipe in self.pipe_sprites:
-            if pipe.right <= 0:
-                pipe.kill()
+            for pipe in self.pipe_sprites:
+                if pipe.right <= 0:
+                    pipe.kill()
 
-        self.pipe_sprites.update()
-        self.bird.update()
+            self.pipe_sprites.update()
+            self.bird.update()
 
-        hit = arcade.check_for_collision_with_list(self.bird, self.pipe_sprites)
-        if any(hit):
-            self.bird.die()
-
+            hit = arcade.check_for_collision_with_list(self.bird, self.pipe_sprites)
+            if any(hit):
+                self.bird.die()
+                self.state = State.GAME_OVER
+        elif self.state == State.GAME_OVER:
+            # self.ui_manager = UIManager()
+            # self.ui_manager.add_ui_element(self.play_again)
+            # arcade.start_render()
+            pass
 
 def run_game():
     os.chdir(os.path.split(os.path.realpath(__file__))[0])
